pyproject.toml
[project]
name = "customer-churn-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=40.1.2",
    "fastapi>=0.128.0",
    "flask>=3.1.2",
    "matplotlib>=3.10.8",
    "mysql-connector-python>=9.5.0",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.40.0",
]
